+++ web/src/app/schedules/temp-sched/TempSchedDialog.tsx
@@ -1,7 +1,6 @@
-import React, { useState, useRef, Suspense } from 'react'
+import React, { useState, useRef, Suspense, useMemo } from 'react'
 import { useMutation, gql } from 'urql'
 import Checkbox from '@mui/material/Checkbox'
-import DialogContentText from '@mui/material/DialogContentText'
 import FormControlLabel from '@mui/material/FormControlLabel'
 import FormHelperText from '@mui/material/FormHelperText'
 import Grid from '@mui/material/Grid'
@@ -14,23 +13,15 @@ import _ from 'lodash'
 import { DateTime, Duration, Interval } from 'luxon'
 import { fieldErrors, nonFieldErrors } from '../../util/errutil'
 import FormDialog from '../../dialogs/FormDialog'
-import {
-  contentText,
-  inferDuration,
-  Shift,
-  TempSchedValue,
-} from './sharedUtils'
-import { FormContainer, FormField } from '../../forms'
-import TempSchedAddNewShift from './TempSchedAddNewShift'
-import { isISOAfter, parseInterval } from '../../util/shifts'
+import { inferDuration, Shift, TempSchedValue } from './sharedUtils'
+import { FormContainer } from '../../forms'
+import { parseInterval } from '../../util/shifts'
 import { useScheduleTZ } from '../useScheduleTZ'
 import TempSchedShiftsList from './TempSchedShiftsList'
-import { ISODateTimePicker } from '../../util/ISOPickers'
 import { getCoverageGapItems } from './shiftsListUtil'
-import { fmtLocal } from '../../util/timeFormat'
 import { ensureInterval } from '../timeUtil'
 import TempSchedConfirmation from './TempSchedConfirmation'
-import { TextField, MenuItem, Divider } from '@mui/material'
+import TempSchedForm from './TempSchedForm'
 
 const mutation = gql`
   mutation ($input: SetTemporaryScheduleInput!) {
@@ -43,10 +34,6 @@ function shiftEquals(a: Shift, b: Shift): boolean {
 }
 
 const useStyles = makeStyles((theme: Theme) => ({
-  contentText,
-  avatar: {
-    backgroundColor: theme.palette.primary.main,
-  },
   formContainer: {
     height: '100%',
   },
@@ -62,13 +49,6 @@ const useStyles = makeStyles((theme: Theme) => ({
       paddingLeft: '1rem',
     },
     overflow: 'hidden',
-  },
-  sticky: {
-    position: 'sticky',
-    top: 0,
-  },
-  tzNote: {
-    fontStyle: 'italic',
   },
 }))
 
@@ -90,7 +70,7 @@ const clampForward = (nowISO: string, iso: string): string => {
   return iso
 }
 
-interface DurationValues {
+export interface DurationValues {
   dur: number
   ivl: string
 }
@@ -102,9 +82,11 @@ export default function TempSchedDialog({
   edit = false,
 }: TempScheduleDialogProps): JSX.Element {
   const classes = useStyles()
-  const { q, zone, isLocalZone } = useScheduleTZ(scheduleID)
-  const [now] = useState(DateTime.utc().startOf('minute').toISO())
+  const { q, zone } = useScheduleTZ(scheduleID)
+  const now = useMemo(() => DateTime.utc().startOf('minute').toISO(), [])
   const [showForm, setShowForm] = useState(false)
+
+  const [{ fetching, error }, commit] = useMutation(mutation)
 
   let defaultShiftDur = {} as DurationValues
 
@@ -126,7 +108,7 @@ export default function TempSchedDialog({
 
   const [durValues, setDurValues] = useState<DurationValues>(defaultShiftDur)
 
-  const [value, setValue] = useState({
+  const [value, setValue] = useState<TempSchedValue>({
     start: clampForward(now, _value.start),
     end: _value.end,
     clearStart: edit ? _value.start : null,
@@ -156,15 +138,6 @@ export default function TempSchedDialog({
   const [allowNoCoverage, setAllowNoCoverage] = useState(false)
   const [submitAttempt, setSubmitAttempt] = useState(false) // helps with error messaging on step 1
   const [submitSuccess, setSubmitSuccess] = useState(false)
-
-  const [{ fetching, error }, commit] = useMutation(mutation)
-
-  function validate(): Error | null {
-    if (isISOAfter(value.start, value.end)) {
-      return new Error('Start date/time cannot be after end date/time.')
-    }
-    return null
-  }
 
   const hasInvalidShift = (() => {
     if (q.loading) return false
@@ -287,7 +260,7 @@ export default function TempSchedDialog({
 
   return (
     <FormDialog
-      fullHeight
+      fullScreen
       maxWidth='lg'
       title={edit ? 'Edit a Temporary Schedule' : 'Define a Temporary Schedule'}
       onClose={onClose}
@@ -319,133 +292,17 @@ export default function TempSchedDialog({
                 justifyContent='space-between'
               >
                 {/* left pane */}
-                <Grid
-                  item
-                  xs={12}
-                  md={6}
-                  container
-                  alignContent='flex-start'
-                  spacing={2}
-                >
-                  <Grid item xs={12}>
-                    <DialogContentText className={classes.contentText}>
-                      The schedule will be exactly as configured here for the
-                      entire duration (ignoring all assignments and overrides).
-                    </DialogContentText>
-                  </Grid>
-
-                  <Grid item xs={12}>
-                    <Typography
-                      color='textSecondary'
-                      className={classes.tzNote}
-                    >
-                      Times shown in schedule timezone ({zone})
-                    </Typography>
-                  </Grid>
-
-                  <Grid item xs={12} md={6}>
-                    <FormField
-                      fullWidth
-                      component={ISODateTimePicker}
-                      required
-                      name='start'
-                      label='Schedule Start'
-                      min={now}
-                      max={DateTime.fromISO(now, { zone })
-                        .plus({ year: 1 })
-                        .toISO()}
-                      softMax={value.end}
-                      softMaxLabel='Must be before end time.'
-                      softMin={DateTime.fromISO(value.end)
-                        .plus({ month: -3 })
-                        .toISO()}
-                      softMinLabel='Must be within 3 months of end time.'
-                      validate={() => validate()}
-                      timeZone={zone}
-                      disabled={q.loading}
-                      hint={isLocalZone ? '' : fmtLocal(value.start)}
-                    />
-                  </Grid>
-                  <Grid item xs={12} md={6}>
-                    <FormField
-                      fullWidth
-                      component={ISODateTimePicker}
-                      required
-                      name='end'
-                      label='Schedule End'
-                      min={now}
-                      softMin={value.start}
-                      softMinLabel='Must be after start time.'
-                      softMax={DateTime.fromISO(value.start)
-                        .plus({ month: 3 })
-                        .toISO()}
-                      softMaxLabel='Must be within 3 months of start time.'
-                      validate={() => validate()}
-                      timeZone={zone}
-                      disabled={q.loading}
-                      hint={isLocalZone ? '' : fmtLocal(value.end)}
-                    />
-                  </Grid>
-
-                  <FormContainer
-                    value={durValues}
-                    onChange={(newValue: DurationValues) => {
-                      setDurValues({ ...durValues, ...newValue })
-                      setValue({
-                        ...value,
-                        shiftDur: Duration.fromObject({
-                          [newValue.ivl]: newValue.dur,
-                        }),
-                      })
-                    }}
-                  >
-                    <Grid item xs={12} md={6}>
-                      <FormField
-                        fullWidth
-                        component={TextField}
-                        type='number'
-                        name='dur'
-                        min={1}
-                        label='Shift Duration'
-                        validate={() => validate()}
-                        disabled={q.loading}
-                      />
-                    </Grid>
-                    <Grid item xs={12} md={6}>
-                      <FormField
-                        fullWidth
-                        component={TextField}
-                        name='ivl'
-                        select
-                        label='Shift Interval'
-                        validate={() => validate()}
-                        disabled={q.loading}
-                      >
-                        <MenuItem value='hours'>Hour</MenuItem>
-                        <MenuItem value='days'>Day</MenuItem>
-                        <MenuItem value='weeks'>Week</MenuItem>
-                      </FormField>
-                    </Grid>
-                  </FormContainer>
-
-                  <Grid item xs={12}>
-                    <Divider />
-                  </Grid>
-
-                  <Grid item xs={12} className={classes.sticky}>
-                    <TempSchedAddNewShift
-                      value={value}
-                      onChange={(shifts: Shift[]) =>
-                        setValue({ ...value, shifts })
-                      }
-                      scheduleID={scheduleID}
-                      showForm={showForm}
-                      setShowForm={setShowForm}
-                      shift={shift}
-                      setShift={setShift}
-                    />
-                  </Grid>
-                </Grid>
+                <TempSchedForm
+                  scheduleID={scheduleID}
+                  duration={durValues}
+                  setDuration={setDurValues}
+                  value={value}
+                  setValue={setValue}
+                  showForm={showForm}
+                  setShowForm={setShowForm}
+                  shift={shift}
+                  setShift={setShift}
+                />
 
                 {/* right pane */}
                 <Grid
